requirements.txt
Django==1.5.1
django-filter==0.6
django-grappelli==2.4.4
djangorestframework==2.2.7
wsgiref==0.1.2
